requirements.txt
Flask==3.1.1
pymongo==4.13.0
python-dotenv==1.1.0
